requirements.txt
flask==2.0.2
Flask_Cors==3.0.10
Flask-Caching==1.10.1
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
marshmallow-sqlalchemy==0.26.1
Passlib==1.7.4
python-dotenv
PyMySQL==1.0.2
pytest==7.1.2
pytest-flask==1.2.0
requests==2.26.0
SQLAlchemy==1.4.26
debugpy
werkzeug==2.0.3
